+++ stubs/SilverStripe/Control/Controller.php
@@ -9,5 +9,17 @@ if (class_exists('SilverStripe\Control\Controller')) {
 
 class Controller
 {
+    /**
+     * @return bool
+     */
+    public static function has_curr()
+    {
+    }
 
+    /**
+     * @return Controller|null
+     */
+    public static function curr()
+    {
+    }
 }