+++ tests/Control/ReplaceHasCurrWithCurrRector/config/configured_rule.php
@@ -1,0 +1,8 @@
+<?php
+
+use Netwerkstatt\SilverstripeRector\Rector\Control\ReplaceHasCurrWithCurrRector;
+use Rector\Config\RectorConfig;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->rule(ReplaceHasCurrWithCurrRector::class);
+};